Finished analysis of RNA/DNA bias between subpools
Don't think these plots are that helpful but made plots looking at the distribution of expression about the backgrounds in both conditions. Made a plot using the ratio of summed reads RNA/summed reads DNA to compare with plot using the sum of RNA. Can remove this later.

Copied subpool manipulations using the DNA ratio to the section following 'Normalizing RNA reads to DNA' to keep this analysis in one section.