Median analysis barebones
Updated the median scripts to be like the integrated analysis. Added a pseduocount of 1 for all reads but set a minimum of 2 reads in DNA, need to change this to 4 like in integrated. Set a minimum BC # per variant of 8 and plotted median vs. sum as well as replicates using median values.